pyproject.toml
[project]
name = "faq-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=0.3.21",
    "langchain-openai>=0.3.11",
    "python-dotenv>=1.1.0",
]
